VA-Helper/requirements.txt
importlib-metadata==8.0.0
jaraco.collections==5.1.0
langchain-community==0.3.27
ollama==0.5.1
pdf2image==1.17.0
pi-heif==1.0.0
pip-chill==1.0.3
platformdirs==4.2.2
streamlit==1.47.1
tomli==2.0.1
unstructured==0.18.11
unstructured-inference==1.0.5
